# docs/swagger.yaml
definitions:
  models.AirdropAddressArray:
    properties:
      address:
        items:
          type: string
        type: array
      amount:
        items:
          type: string
        type: array
    type: object
  models.AirdropAdminSignData:
    properties:
      account:
        description: personal_sign address,0x
        type: string
      airdropId:
        description: airdrop2/airdropDownload:need AirdropId
        type: integer
      array:
        $ref: '#/definitions/models.AirdropAddressArray'
        description: airdrop2:need Array
      chainId:
        description: airdrop1:need ChainId
        type: integer
      daoAddress:
        description: airdrop1:need DaoAddress
        type: string
      message:
        description: '{"expired":1244,"root": "","type":"airdrop1/airdrop2/airdropDownload"}'
        type: string
      signature:
        description: personal_sign sign result,no 0x
        type: string
    type: object
  models.CollectInfo:
    properties:
      name:
        type: string
      required:
        type: boolean
    type: object
  models.CreateAirdropParam:
    properties:
      airdropEndTime:
        type: integer
      airdropStartTime:
        type: integer
      collectInformation:
        items:
          $ref: '#/definitions/models.CollectInfo'
        type: array
      description:
        type: string
      endTime:
        type: integer
      maxAirdropAmount:
        type: string
      sign:
        $ref: '#/definitions/models.AirdropAdminSignData'
      startTime:
        type: integer
      title:
        type: string
      tokenAddress:
        type: string
      tokenChainId:
        type: integer
    type: object
...
  models.ResAirdropId:
    properties:
      airdropId:
        type: integer
      signature:
        type: string
    type: object
  models.ResAirdropInfo:
    properties:
      addressNum:
        type: integer
      airdropEndTime:
        type: integer
      airdropStartTime:
        type: integer
      chainId:
        type: integer
      collect:
        items:
          $ref: '#/definitions/models.CollectInfo'
        type: array
      collectCount:
        type: integer
      creator:
        type: string
      daoAddress:
        type: string
      description:
        type: string
      endTime:
        type: integer
      startTime:
        type: integer
      title:
        type: string
      tokenAddress:
        type: string
      tokenChainId:
        type: integer
    type: object
...
  models.ResProof:
    properties:
      airdropNumber:
        type: integer
      airdropTotalAmount:
        type: string
      amount:
        type: string
      index:
        type: integer
      proof:
        items:
          type: string
        type: array
      title:
        type: string
    type: object
...
  models.ResTreeRoot:
    properties:
      root:
        type: string
    type: object
  models.ResUploadAddressList:
    properties:
      address:
        type: string
      amount:
        type: string
    type: object
...
  models.Response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
...
  models.UserInformationParam:
    properties:
      account:
        type: string
      airdropId:
        type: integer
      userSubmit:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /stpdao/v2/airdrop/address:
    post:
      description: save airdrop address
      parameters:
      - description: request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.AirdropAdminSignData'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ResTreeRoot'
      summary: save airdrop address
      tags:
      - Airdrop
  /stpdao/v2/airdrop/address/list:
    get:
      description: upload airdrop address list
      parameters:
      - description: airdropId
        in: query
        name: airdropId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ResUploadAddressList'
      summary: upload airdrop address list
      tags:
      - Airdrop
  /stpdao/v2/airdrop/collect:
    get:
      description: airdrop need collect information
      parameters:
      - description: id
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ResAirdropInfo'
      summary: airdrop need collect information
      tags:
      - Airdrop
  /stpdao/v2/airdrop/create:
    post:
      description: create airdrop
      parameters:
      - description: request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CreateAirdropParam'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ResAirdropId'
      summary: create airdrop
      tags:
      - Airdrop
  /stpdao/v2/airdrop/proof:
    get:
      description: claim airdrop address
      parameters:
      - description: address
        in: query
        name: address
        required: true
        type: string
      - description: id
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ResProof'
      summary: claim airdrop address
      tags:
      - Airdrop
  /stpdao/v2/airdrop/save/user:
    post:
      description: airdrop collect user information
      parameters:
      - description: request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.UserInformationParam'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
      summary: airdrop collect user information
      tags:
      - Airdrop
  /stpdao/v2/airdrop/user/download:
    post:
      description: airdrop download user information
      parameters:
      - description: request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.AirdropAdminSignData'
      produces:
      - application/json
      responses: {}
      summary: airdrop download user information
      tags:
      - Airdrop
...
swagger: "2.0"